# 09-aws-interop/infra/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pattern 9 simulated AWS/Bedrock REST capability",
    "description": "[owner:09-cross-cloud-protocol-gateway] Industry-neutral deterministic capability. AWS Lambda and Amazon Bedrock are simulated.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://pattern9-backend.example.invalid"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Healthy simulated backend",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "operationId": "listCapabilities",
        "summary": "List deterministic quote capabilities",
        "responses": {
          "200": {
            "description": "Supported capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/quotes": {
      "post": {
        "operationId": "createQuote",
        "summary": "Create a deterministic capability quote",
        "description": "Returns the same quote and correlation IDs for the same normalized request.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Deterministic simulated AWS/Bedrock quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "QuoteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "capability",
          "serviceLevel",
          "units",
          "termMonths"
        ],
        "properties": {
          "capability": {
            "type": "string",
            "enum": [
              "document-processing",
              "data-analysis",
              "workflow-automation"
            ]
          },
          "serviceLevel": {
            "type": "string",
            "enum": [
              "standard",
              "priority",
              "critical"
            ]
          },
          "units": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "termMonths": {
            "type": "integer",
            "enum": [
              1,
              12,
              24,
              36
            ]
          }
        }
      },
      "QuoteResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/QuoteRequest"
          },
          {
            "type": "object",
            "required": [
              "quoteId",
              "correlationId",
              "currency",
              "unitPrice",
              "totalPrice",
              "simulation",
              "simulatedBackend"
            ],
            "properties": {
              "quoteId": {
                "type": "string"
              },
              "correlationId": {
                "type": "string"
              },
              "currency": {
                "type": "string",
                "enum": [
                  "USD"
                ]
              },
              "unitPrice": {
                "type": "string"
              },
              "totalPrice": {
                "type": "string"
              },
              "validForMinutes": {
                "type": "integer",
                "enum": [
                  15
                ]
              },
              "simulation": {
                "type": "boolean",
                "enum": [
                  true
                ]
              },
              "simulatedBackend": {
                "type": "string",
                "enum": [
                  "AWS Lambda / Amazon Bedrock (simulated)"
                ]
              }
            }
          }
        ]
      }
    }
  }
}